Update README.md
shoppingCart is an online shopping application that helps its users to purchase mobiles and tablet devices. This application allows users to log in for purchasing mobiles and tablet devices. Users can search for a product, sort the products list based on rating or price. Users can select the products and add them to the cart.  Once the selection is done, users can go to the cart page for payment. User can change the quantity of the products selected for purchase, Checkout for closing the purchase and logout of the appliation.